app.py

import streamlit as st
import random
import stripe
import os
from dotenv import load_dotenv

load_dotenv()

stripe.api_key = os.getenv("STRIPE_SECRET_KEY")

# Sample placeholder for AI Dating App
st.title("AIBonds: AI Agent Dating App - Premium Edition")
st.write("Generate love, poetry, and more. Upgrade to Premium for full features!")

# Stripe payment simulation button
if st.button("Go Premium"):
    st.success("Redirecting to Stripe checkout... (Simulated)")
